refactor(Spec): create `Contracts` namespace for all spec related interfaces
