+++ src/Utils/SpecificationWalker.php
@@ -6,7 +6,7 @@
 
 namespace OpenApi\Utils;
 
-use OpenApi\AttributeInterface;
+use OpenApi\Contracts\AttributeInterface;
 use OpenApi\Spec as OA;
 use OpenApi\Specification;
